requirements.txt
Django>=6.0
djangorestframework>=3.16
Pillow>=12.0
psycopg2-binary>=2.9
python-dotenv>=1.0
